requirements.txt
pytest==7.1.0
python-dotenv==0.19.2
requests==2.27.1
sendgrid==6.9.7
black==22.3.0
google-cloud-secret-manager==2.9.2
numpy==1.22.3
pymongo==4.0.2
jsonschema==4.4.0
responses==0.20.0
mongomock==4.0.0
pylint==2.13.5